fix: fix issue with rmDir failing

+++ internal/docs/scripts/generate-markdown.js
@@ -16,7 +16,7 @@ async function main() {
   const cwd = process.cwd();
   const RELATIVE_PKG_ROOT = `${cwd}/../../packages`;
 
-  fs.rmdirSync(`${cwd}/build`, { recursive: true });
+  fs.existsSync(`${cwd}/build`) && fs.rmdirSync(`${cwd}/build`, { recursive: true });
   fs.mkdirSync(`${cwd}/build/docs`, { recursive: true });
 
   fs.copySync('content', `${cwd}/build/docs/content`);
